+++ src/Contracts/ViewMendManagerContract.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace ViewMend\Laravel\Contracts;
 
+use ViewMend\Cron\CronClient;
 use ViewMend\Laravel\ViewMendConnection;
 use ViewMend\SiteTracker\SiteTrackerClient;
 use ViewMend\ViewMend;
@@ -15,4 +16,6 @@ interface ViewMendManagerContract
     public function client(): ViewMend;
 
     public function siteTracker(): SiteTrackerClient;
+
+    public function cron(): CronClient;
 }